+++ src/application.h
@@ -61,7 +61,7 @@ private:
     void createWalletd();
     void subscribeToWalletd();
     void firstRun();
-    void runBuiltinWalletd(const QString& walletFile, bool createNew, bool createLegacy, QByteArray&& keys, QByteArray&& mnemonic);
+    void runBuiltinWalletd(const QString& walletFile, bool createNew, bool createLegacy, bool createHardware, QByteArray&& keys, QByteArray&& mnemonic);
 
 signals:
     void builtinRunSignal();
@@ -86,6 +86,7 @@ public slots:
 
     void createLegacyWallet(QWidget* parent);
     void createWallet(QWidget* parent);
+    void createHWWallet(QWidget* parent);
     void openWallet(QWidget* parent);
     void restoreWalletFromMnemonic(QWidget* parent);
     void remoteWallet(QWidget* parent);